commit_3_finished modal

+++ src/app/todo/todo.component.ts
@@ -1,4 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { ConfirmationModalComponent } from '../confirmation-modal/confirmation-modal.component';
 import { ITodo } from '../interfaces/itodo';
 import { TodoService } from '../services/todo.service';
 
@@ -9,13 +11,25 @@ import { TodoService } from '../services/todo.service';
 })
 export class TodoComponent implements OnInit {
 
-  @Input()todo: ITodo;
-  constructor(private todoService: TodoService) { }
+  @Input() todo: ITodo;
+  constructor(private todoService: TodoService, private modalService: NgbModal) { }
 
   ngOnInit() {
   }
 
-  deleteTodo() {
-    this.todoService.deleteTodo(this.todo);
+  async deleteTodo() {
+    const modal = this.modalService.open(ConfirmationModalComponent);
+    modal.componentInstance.modalInstance = modal;
+
+    let answer = ''
+    try {
+      answer = await modal.result;
+    }
+    catch (error) {
+      console.log(error);
+    }
+    if (answer === 'yes') {
+      this.todoService.deleteTodo(this.todo);
+    }
   }
 }
